third_party/blink/web_tests/platform/win/external/wpt/editing/run/italic_1001-2000-expected.txt

This is a testharness.js-based test.
Found 165 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<dfn>bar</dfn>b]az" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<dfn>bar</dfn>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<dfn>bar</dfn>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<dfn>bar</dfn>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<dfn>bar</dfn>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<em>bar</em>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<em>bar</em>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<em>bar</em>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<em>bar</em>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<var>bar</var>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<var>bar</var>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<var>bar</var>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<var>bar</var>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<address>bar</address>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<cite>bar</cite>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite>bar</cite><span style=\\"font-style:italic\\">baz</span>" but got "foo<cite>bar</cite>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<cite>bar</cite>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<cite>bar</cite><i>baz</i>" but got "foo<cite>bar</cite>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<cite>bar</cite>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<dfn>bar</dfn>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn>bar</dfn><span style=\\"font-style:italic\\">baz</span>" but got "foo<dfn>bar</dfn>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<dfn>bar</dfn>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<dfn>bar</dfn><i>baz</i>" but got "foo<dfn>bar</dfn>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<dfn>bar</dfn>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<em>bar</em>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<em>bar</em>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<em>bar</em>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<em>bar</em>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>bar</i>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>bar</i>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>bar</i>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>bar</i>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<var>bar</var>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var>bar</var><span style=\\"font-style:italic\\">baz</span>" but got "foo<var>bar</var>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<var>bar</var>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<var>bar</var><i>baz</i>" but got "foo<var>bar</var>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<var>bar</var>baz]" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandIndeterm("italic") after
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandState("italic") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<address>bar</address>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<cite>bar</cite>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<cite>bar</cite>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<cite>bar</cite>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<cite>bar</cite>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<dfn>bar</dfn>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<dfn>bar</dfn>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<dfn>bar</dfn>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<dfn>bar</dfn>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<em>bar</em>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<em>bar</em>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<em>bar</em>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<em>bar</em>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<i>bar</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<i>bar</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<i>bar</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<i>bar</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<var>bar</var>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "[foo<var>bar</var>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<var>bar</var>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "[foo<var>bar</var>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: italic\\">[bar]</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: italic\\">[bar]</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: italic\\">[bar]</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: italic\\">[bar]</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: oblique\\">[bar]</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: oblique\\">[bar]</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: oblique\\">[bar]</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: oblique\\">[bar]</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: oblique\\">b[a]r</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<span style=\\"font-style: oblique\\">b[a]r</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: oblique\\">b[a]r</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<span style=\\"font-style: oblique\\">b[a]r</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p><span style=\\"font-style:italic\\">baz</span></p>" but got "<p style=\\"\\">foo</p><p style=\\"\\">bar</p><p style=\\"font-style:italic\\">baz</p>"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p><p><i>baz</i></p>" but got "<p style=\\"\\">foo</p><p style=\\"\\">bar</p><p style=\\"font-style:italic\\">baz</p>"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i>{<p>foo</p><p>bar</p>}<p>baz</p></i>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><span style=\\"font-style:italic\\">foo</span><b>bar</b></p><p><span style=\\"font-style:italic\\">baz</span></p>" but got "<p style=\\"\\"><span style=\\"font-style:italic\\">foo</span><b style=\\"\\">bar</b></p><p style=\\"font-style:italic\\">baz</p>"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><i>foo</i><b>bar</b></p><p><i>baz</i></p>" but got "<p style=\\"\\"><i>foo</i><b style=\\"\\">bar</b></p><p style=\\"font-style:italic\\">baz</p>"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "<i><p>foo[<b>bar</b>}</p><p>baz</p></i>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo [bar <b>baz] qoz</b> quz sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo bar <b>baz [qoz</b> quz] sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo [bar <i>baz] qoz</i> quz sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo [bar <i>baz] qoz</i> quz sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo bar <i>baz [qoz</i> quz] sic" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo bar <i>baz [qoz</i> quz] sic" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "fo[o<i>b]ar</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "fo[o<i>b]ar</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["italic",""]] "foo<i>ba[r</i>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "foo<i>ba[r</i>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<i>bar</i>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<i>bar</i>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<i>bar</i>b]az" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<i>bar</i>b]az" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>b]ar</i>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foob<span style=\\"font-style:italic\\">ar</span>baz" but got "foob<i>ar</i>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>b]ar</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>b]ar</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>b]ar</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>b]ar</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>ba[r</i>]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-style:italic\\">ba</span>rbaz" but got "foo<i>ba</i>rbaz"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>ba[r</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>ba[r</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>ba[r</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>ba[r</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>bar</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo[<i>bar</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>bar</i>]baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo[<i>bar</i>]baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>[bar]</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo<i>[bar]</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>[bar]</i>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo<i>[bar]</i>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<i>bar</i>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "foo{<i>bar</i>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<i>bar</i>}baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "foo{<i>bar</i>}baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "fo[o<span style=font-style:italic>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"font-style:italic\\">obar</span>baz" but got "fo<i>ob</i><span style=\\"font-style:italic\\">ar</span>baz"
[FAIL] [["italic",""]] "fo[o<span style=font-style:italic>b]ar</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["italic",""]] "fo[o<span style=font-style:italic>b]ar</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<span style=\\"font-style:italic\\">o</span><span style=\\"font-style:oblique\\"><span style=\\"font-style:italic\\">b</span>ar</span>baz" but got "fo<span style=\\"font-style:italic\\">ob</span><span style=\\"font-style:oblique\\">ar</span>baz"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<i>o</i><span style=\\"font-style:oblique\\"><i>b</i>ar</span>baz" but got "fo<i>ob</i><span style=\\"font-style:oblique\\">ar</span>baz"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "fo[o<span style=font-style:oblique>b]ar</span>baz" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" queryCommandValue("italic") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" queryCommandValue("italic") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["italic",""]] "<span style=font-style:italic>fo[o</span><span style=font-style:oblique>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<i>fo</i>ob<span style=\\"font-style:oblique\\">ar</span>" but got "<span style=\\"font-style:italic\\">fo</span>ob<span style=\\"font-style:oblique\\">ar</span>"
Harness: the test ran to completion.
